Add guards to ensure delicate operations can only be performed by the logged in user

+++ libs/api/user/feature/src/lib/user.controller.ts
@@ -7,9 +7,15 @@ import {
   NotFoundException,
   Param,
   Patch,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { UserEntity } from '@newbee/api/shared/data-access';
-import { idNotFoundErrorMsg } from '@newbee/api/shared/util';
+import {
+  idNotFoundErrorMsg,
+  unauthorizedErrorMsg,
+  unauthorizedLogMsg,
+  User,
+} from '@newbee/api/shared/util';
 import { UpdateUserDto, UserService } from '@newbee/api/user/data-access';
 import { userVersion } from '@newbee/shared/data-access';
 
@@ -35,21 +41,37 @@ export class UserController {
   @Patch(':id')
   async update(
     @Param('id') id: string,
-    @Body() updateUserDto: UpdateUserDto
+    @Body() updateUserDto: UpdateUserDto,
+    @User() user: UserEntity
   ): Promise<UserEntity> {
     this.logger.log(
       `Update user request received for id: ${id}: ${JSON.stringify(
         updateUserDto
       )}`
     );
-    const user = await this.userService.update(id, updateUserDto);
-    this.logger.log(`Updated user: ${JSON.stringify(user)}`);
-    return user;
+
+    if (id !== user.id) {
+      this.logger.error(unauthorizedLogMsg('update', user.id, id));
+      throw new UnauthorizedException(unauthorizedErrorMsg);
+    }
+
+    const updatedUser = await this.userService.update(id, updateUserDto);
+    this.logger.log(`Updated user: ${JSON.stringify(updatedUser)}`);
+    return updatedUser;
   }
 
   @Delete(':id')
-  async delete(@Param('id') id: string): Promise<void> {
+  async delete(
+    @Param('id') id: string,
+    @User() user: UserEntity
+  ): Promise<void> {
     this.logger.log(`Delete user request received for id: ${id}`);
+
+    if (id !== user.id) {
+      this.logger.error(unauthorizedErrorMsg);
+      throw new UnauthorizedException(unauthorizedErrorMsg);
+    }
+
     await this.userService.delete(id);
     this.logger.log(`Deleted user for id: ${id}`);
   }
